refactor(table): move interfaces and utility functions to separate files
This commit refactors the PriceTable component by moving the TypeScript interfaces and utility functions (transformPriceData and createColumns) to separate files (table.props.ts and lib/index.ts). This improves code organization, maintainability, and reusability across the codebase.
